+++ lib/simple_auth/compat.rb
@@ -1,2 +1,0 @@
-require "simple_auth/compat/config"
-require "simple_auth/compat/active_record"