+++ lib/simple_auth/exceptions.rb
@@ -1,4 +1,0 @@
-module SimpleAuth
-  class RecordNotFound < StandardError; end
-  class NotAuthorized < StandardError; end
-end