+++ lib/simple_auth/generator.rb
@@ -1,8 +1,8 @@
-require "rails/generators/base"
+require "rails/generators"
 
 module SimpleAuth
   class InstallGenerator < ::Rails::Generators::Base
-    source_root File.dirname(__FILE__) + "/../../templates"
+    source_root "#{__dir__}/templates/install"
 
     def copy_initializer_file
       copy_file "initializer.rb", "config/initializers/simple_auth.rb"