+++ spec/support/app/models/person.rb
@@ -1,4 +1,0 @@
-class Person < ActiveRecord::Base
-  self.table_name = "users"
-  authentication validations: false
-end