+++ test/support/dummy/app/controllers/pages_controller.rb
@@ -1,0 +1,7 @@
+class PagesController < ApplicationController
+  before_action :require_logged_user
+
+  def index
+    head :ok
+  end
+end