+++ test/support/dummy/config/routes.rb
@@ -1,0 +1,23 @@
+Rails.application.routes.draw do
+  get "/dashboard", to: "dashboard#index"
+  get "/admin/dashboard", to: "admin/dashboard#index"
+  get "/login", to: "sessions#new"
+
+  controller :dashboard do
+    get :log_in
+    get :not_logged
+  end
+
+  controller :pages do
+    get :index
+  end
+
+  namespace :admin do
+    controller :dashboard do
+      get :index
+      get :log_in_as_admin
+      get :log_in_as_user
+      get :log_in_with_admin_flag
+    end
+  end
+end